extract popover to exported component
