+++ resources/tentative/complex-static-html/src/big-dom-generator/components/top-bar.jsx
@@ -1,5 +1,6 @@
 import { SearchArea } from "./search-area";
 import { ActionButton, ActionGroup } from "./action-group";
+import { PopOver } from "./pop-over";
 
 import ProfileIcon from "./../assets/Smock_RealTimeCustomerProfile_18_N.svg";
 import SettingsIcon from "./../assets/Smock_Settings_18_N.svg";
@@ -8,6 +9,9 @@ import BellIcon from "./../assets/Smock_Bell_18_N.svg";
 import SpeedometerLogo from "./../assets/logo.png";
 
 export const TopBar = () => {
+    const NUM_OF_NOTIFICATIONS = 10;
+    const NUM_OF_SETTINGS = 5;
+    const NUM_OF_PROFILE_OPTIONS = 3;
     return (
         <div className="ui top-bar">
             <img className="ui" src={SpeedometerLogo} alt="Speedometer Logo" height={40} />
@@ -18,8 +22,11 @@ export const TopBar = () => {
             <div className="ui top-bar-right">
                 <ActionGroup>
                     <ActionButton Icon={BellIcon} quite={false} />
+                    <PopOver numOptions={NUM_OF_NOTIFICATIONS} />
                     <ActionButton Icon={SettingsIcon} quite={false} />
+                    <PopOver numOptions={NUM_OF_SETTINGS} />
                     <ActionButton Icon={ProfileIcon} quite={false} />
+                    <PopOver numOptions={NUM_OF_PROFILE_OPTIONS} />
                 </ActionGroup>
             </div>
         </div>
